refactor ErgoTool and add `address` command
