+++ java-examples/build.gradle.kts
@@ -26,7 +26,7 @@ repositories {
 
 dependencies {
     testImplementation("junit", "junit", "4.12")
-    implementation("org.ergoplatform", "ergo-appkit_2.12", "sandboxed-4c5664c8-SNAPSHOT", "compile")
+    implementation("org.ergoplatform", "ergo-appkit_2.12", "sandboxed-4da679af-SNAPSHOT", "compile")
 //    implementation("org.ergoplatform", "ergo-appkit_2.12", "3.1.0", "compile")
     implementation("org.graalvm.sdk", "graal-sdk", "19.2.1", "compile")
 }
