+++ scala-examples/build.gradle.kts
@@ -25,10 +25,13 @@ repositories {
 
 dependencies {
     implementation("org.scala-lang:scala-library:2.12.8")
-    implementation("org.ergoplatform", "ergo-appkit_2.12", "sandboxed-4c5664c8-SNAPSHOT", "compile")
+    implementation("org.ergoplatform", "ergo-appkit_2.12", "sandboxed-4da679af-SNAPSHOT", "compile")
     implementation("org.graalvm.sdk", "graal-sdk", "19.2.1", "compile")
     implementation("com.squareup.okhttp3:mockwebserver:3.12.0")
-    testImplementation("junit", "junit", "4.12")
+//    testImplementation("junit", "junit", "4.12")
+    testImplementation("org.scalatest", "scalatest_2.12", "3.0.8")
+    testImplementation("org.scalacheck", "scalacheck_2.12", "1.14.2")
+
 }
 
 java {
